requirements.txt
openai>=1.40.0
python-slugify>=8.0.4
PyYAML>=6.0.1